+++ .storybook/manager.js
@@ -1,0 +1,6 @@
+import { addons } from '@storybook/manager-api';
+import depositsTheme from './DepositsTheme';
+
+addons.setConfig({
+    theme: depositsTheme,
+});