+++ storybook-static/assets/Autolayout.stories-d50ec54d.js
@@ -1,0 +1,10 @@
+import{j as o}from"./jsx-runtime-94f6e698.js";import{A as a}from"./AutoLayout-7fd52157.js";import{B as r}from"./Box-30be3602.js";import"./index-8db94870.js";import"./_commonjsHelpers-042e6b4d.js";import"./index-58d3fd43.js";const c={title:"Layouts/Auto layout",component:a,argTypes:{}},e=i=>o.jsxs(a,{...i,children:[o.jsx(r,{background:"#0CB9E9",borderRadius:"6px",height:"48px",width:"48px"}),o.jsx(r,{background:"#0CB9E9",borderRadius:"6px",height:"48px",width:"48px"}),o.jsx(r,{background:"#0CB9E9",borderRadius:"6px",height:"48px",width:"48px"}),o.jsx(r,{background:"#0CB9E9",borderRadius:"6px",height:"48px",width:"48px"}),o.jsx(r,{background:"#0CB9E9",borderRadius:"6px",height:"48px",width:"48px"}),o.jsx(r,{background:"#0CB9E9",borderRadius:"6px",height:"48px",width:"48px"}),o.jsx(r,{background:"#0CB9E9",borderRadius:"6px",height:"48px",width:"48px"})]}),t=e.bind({});var d,x,p;t.parameters={...t.parameters,docs:{...(d=t.parameters)==null?void 0:d.docs,source:{originalSource:`args => <AutoLayout {...args}>
+    <Box background="#0CB9E9" borderRadius="6px" height="48px" width="48px" />
+    <Box background="#0CB9E9" borderRadius="6px" height="48px" width="48px" />
+    <Box background="#0CB9E9" borderRadius="6px" height="48px" width="48px" />
+    <Box background="#0CB9E9" borderRadius="6px" height="48px" width="48px" />
+    <Box background="#0CB9E9" borderRadius="6px" height="48px" width="48px" />
+    <Box background="#0CB9E9" borderRadius="6px" height="48px" width="48px" />
+    <Box background="#0CB9E9" borderRadius="6px" height="48px" width="48px" />
+  </AutoLayout>`,...(p=(x=t.parameters)==null?void 0:x.docs)==null?void 0:p.source}}};const B=["Default"];export{t as Default,B as __namedExportsOrder,c as default};
+//# sourceMappingURL=Autolayout.stories-d50ec54d.js.map